+++ Assets/Scripts/ConfigItems/TextInputFieldConfigItem.cs
@@ -8,12 +8,16 @@ namespace LethalConfig.ConfigItems
     public sealed class TextInputFieldConfigItem : BaseValueConfigItem<string>
     {
         internal int CharacterLimit { get; private set; }
+        internal int NumberOfLines { get; private set; }
+        internal bool TrimText { get; private set; }
 
         public TextInputFieldConfigItem(ConfigEntry<string> configEntry) : this(configEntry, true) { }
         public TextInputFieldConfigItem(ConfigEntry<string> configEntry, bool requiresRestart) : this(configEntry, GetDefaultOptions(configEntry, requiresRestart)) { }
         public TextInputFieldConfigItem(ConfigEntry<string> configEntry, TextInputFieldOptions options) : base(configEntry, options)
         {
             CharacterLimit = options.CharacterLimit;
+            NumberOfLines = options.NumberOfLines;
+            TrimText = options.TrimText;
         }
 
         internal override GameObject CreateGameObjectForConfig()
@@ -26,6 +30,7 @@ namespace LethalConfig.ConfigItems
             return new()
             {
                 CharacterLimit = 0,
+                NumberOfLines = 1,
                 RequiresRestart = requiresRestart
             };
         }