+++ Assets/Scripts/MonoBehaviours/Components/TextInputFieldController.cs
@@ -1,7 +1,11 @@
 using LethalConfig.ConfigItems;
 using LethalConfig.MonoBehaviours.Managers;
+using LethalConfig.Utils;
+using System.Collections;
 using TMPro;
 using UnityEngine;
+using UnityEngine.InputSystem;
+using UnityEngine.UI;
 
 namespace LethalConfig.MonoBehaviours.Components
 {
@@ -9,15 +13,45 @@ namespace LethalConfig.MonoBehaviours.Components
     {
         public TMP_InputField textInputField;
 
+        private LayoutElement layoutElement;
+
+        protected override void Awake()
+        {
+            base.Awake();
+
+            layoutElement = GetComponent<LayoutElement>();
+        }
+
+        private void Update()
+        {
+            if (textInputField.isFocused && textInputField.lineType != TMP_InputField.LineType.SingleLine)
+            {
+                var isShiftPressed = Keyboard.current.shiftKey.isPressed;
+                textInputField.lineType = isShiftPressed ? TMP_InputField.LineType.MultiLineNewline : TMP_InputField.LineType.MultiLineSubmit;
+            }
+        }
+
         protected override void OnSetConfigItem()
         {
-            textInputField.SetTextWithoutNotify(ConfigItem.CurrentValue);
+            textInputField.text = ConfigItem.CurrentValue;
+            var lines = Mathf.Clamp(ConfigItem.NumberOfLines <= 0 ? 4 : ConfigItem.NumberOfLines, 1, 4);
+            var height = (float)(16 + (lines * 19));
+            layoutElement.minHeight = height;
+            layoutElement.preferredHeight = height;
+            textInputField.lineType = lines == 1 ? TMP_InputField.LineType.SingleLine : TMP_InputField.LineType.MultiLineSubmit;
+            textInputField.lineLimit = ConfigItem.NumberOfLines;
             UpdateAppearance();
         }
 
         public void OnInputFieldEndEdit(string value)
         {
-            ConfigItem.CurrentValue = value;
+            var caretPosition = textInputField.caretPosition;
+            if (ConfigItem.NumberOfLines != 1)
+            {
+                StartCoroutine(RemoveNewLineFromSubmitDelayed(caretPosition));
+            }
+
+            ConfigItem.CurrentValue = ConfigItem.TrimText ? value.Trim() : value;
             UpdateAppearance();
             ConfigMenuManager.Instance.menuAudio.PlayChangeValueSFX();
         }
@@ -25,8 +59,18 @@ namespace LethalConfig.MonoBehaviours.Components
         public override void UpdateAppearance()
         {
             base.UpdateAppearance();
-            textInputField.SetTextWithoutNotify(ConfigItem.CurrentValue);
+            textInputField.text = ConfigItem.CurrentValue;
             textInputField.textComponent.rectTransform.localPosition = Vector3.zero;
+        }
+
+        // This is a workaround because the TMP_InputField adds a line break when pressing enter
+        // even with the lineType set to MultiLineSubmit. However, it only adds this line AFTER LateUpdate
+        // but before rendering. So this is the only way i could find to remove this additional linebreak it adds to the field.
+
+        private IEnumerator RemoveNewLineFromSubmitDelayed(int caretPosition)
+        {
+            yield return null;
+            textInputField.text = textInputField.text.Remove(Mathf.Clamp(caretPosition, 0, textInputField.text.Length - 1), 1);
         }
     } 
 }